Remove EXE file from codebase and move to release section; update English version from Chinese and add Chinese as a supported language
