+++ language/zh.ts
@@ -1,0 +1,429 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE TS>
+<TS version="2.1" language="zh_CN" sourcelanguage="en_GB">
+<context>
+    <name>AboutDialog</name>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="357"/>
+        <source>About PythonEXE Maker</source>
+        <translation>关于 PythonEXE Maker</translation>
+    </message>
+</context>
+<context>
+    <name>DropArea</name>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="286"/>
+        <source>Drag in .py file</source>
+        <translation>拖入 .py 文件</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="313"/>
+        <source>Warning</source>
+        <translation>警告</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="313"/>
+        <source>Please drag and drop the Python file (.py) into the window.</source>
+        <translation>请拖放 Python 文件 (.py) 到窗口中。</translation>
+    </message>
+</context>
+<context>
+    <name>LogViewerDialog</name>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="394"/>
+        <source>Log File</source>
+        <translation>查看日志文件</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="410"/>
+        <source>Unable to read the log file: {e}</source>
+        <translation>无法读取日志文件: {e}</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="412"/>
+        <source>The log file does not exist.</source>
+        <translation>日志文件不存在。</translation>
+    </message>
+</context>
+<context>
+    <name>MainWindow</name>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="420"/>
+        <source>PythonEXE Maker</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="459"/>
+        <source>Script management</source>
+        <translation>脚本管理</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="468"/>
+        <source>Browse Files</source>
+        <translation>浏览文件</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="469"/>
+        <source>Click to select the Python file to be converted. Multiple selections are available.</source>
+        <translation>点击选择要转换的 Python 文件，可多选。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="480"/>
+        <source>List of selected Python scripts, which can be removed by double-clicking.</source>
+        <translation>已选择的 Python 脚本列表，双击可移除。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="489"/>
+        <source>Conversion Task Progress</source>
+        <translation>转换任务进度</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="503"/>
+        <source>Task Management</source>
+        <translation>任务管理</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="577"/>
+        <source>Log</source>
+        <translation>日志</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="539"/>
+        <source>Icon file not found: {icon_path}</source>
+        <translation>图标文件未找到: {icon_path}</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="545"/>
+        <source>Document</source>
+        <translation type="obsolete">文件</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="546"/>
+        <source>Quit</source>
+        <translation>退出</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="551"/>
+        <source>Help</source>
+        <translation>帮助</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="552"/>
+        <source>Instructions for use</source>
+        <translation>使用说明</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="556"/>
+        <source>About</source>
+        <translation>关于</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="560"/>
+        <source>Open Source Address</source>
+        <translation>开源地址</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="564"/>
+        <source>Official Website</source>
+        <translation>官方网站</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="568"/>
+        <source>LINUXDO Forum Homepage</source>
+        <translation>LINUXDO 论坛主页</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="572"/>
+        <source>Invite Developers to Coffee</source>
+        <translation>请开发者喝咖啡</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="578"/>
+        <source>View log files</source>
+        <translation>查看日志文件</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="590"/>
+        <source>GUI Model</source>
+        <translation>GUI 模式</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="590"/>
+        <source>Command Line Mode</source>
+        <translation>命令行模式</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="591"/>
+        <source>Select whether the generated EXE will be with a console (command line mode) or without a console (GUI mode).</source>
+        <translation>选择生成的 EXE 是带控制台（命令行模式）还是不带控制台（GUI 模式）。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="597"/>
+        <source>Output Directory:</source>
+        <translation>输出目录:</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="599"/>
+        <source>By default, it is in the same directory as the source file.</source>
+        <translation>默认与源文件同目录</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="600"/>
+        <source>Set the output directory for generated EXE files. If empty, it will be placed in the same directory as the source file by default.</source>
+        <translation>设置生成 EXE 文件的输出目录。为空则默认放在源文件同目录下。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="627"/>
+        <source>Browse</source>
+        <translation>浏览</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="603"/>
+        <source>Select the output directory.</source>
+        <translation>选择输出目录。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="615"/>
+        <source>EXE Name:</source>
+        <translation>EXE 名称:</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="617"/>
+        <source>Defaults to the same name as the source file</source>
+        <translation>默认与源文件同名</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="618"/>
+        <source>Set the generated EXE file name.</source>
+        <translation>设置生成的 EXE 文件名称。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="622"/>
+        <source>Icon File:</source>
+        <translation>图标文件:</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="624"/>
+        <source>Optional, supports .png and .ico</source>
+        <translation>可选，支持 .png 和 .ico</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="625"/>
+        <source>Select an icon file to use with the EXE. If it is PNG, it will be automatically converted to ICO.</source>
+        <translation>选择一个图标文件用于 EXE。若为 PNG，将自动转换为 ICO。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="628"/>
+        <source>Select the icon file.</source>
+        <translation>选择图标文件。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="636"/>
+        <source>File Version:</source>
+        <translation>文件版本:</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="639"/>
+        <source>Set the version number of the EXE file (X.X.X.X).</source>
+        <translation>设置 EXE 文件的版本号（X.X.X.X）。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="643"/>
+        <source>Copyright Information:</source>
+        <translation>版权信息:</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="645"/>
+        <source>Set the copyright information of the EXE file.</source>
+        <translation>设置 EXE 文件的版权信息。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="653"/>
+        <source>Advanced Settings</source>
+        <translation>高级设置</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="657"/>
+        <source>Additional Modules:</source>
+        <translation>额外模块:</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="659"/>
+        <source>Hide imported modules, separated by commas</source>
+        <translation type="obsolete">隐藏导入的模块，逗号分隔</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="660"/>
+        <source>Enter the name of the module that needs to be hidden (separate multiple with commas).</source>
+        <translation type="obsolete">输入需要隐藏导入的模块名称（多个用逗号分隔）。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="666"/>
+        <source>Additional Parameters:</source>
+        <translation>附加参数:</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="668"/>
+        <source>For example：--add-data &apos;data.txt;.&apos;</source>
+        <translation>例如：--add-data &apos;data.txt;.&apos;</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="669"/>
+        <source>Enter custom PyInstaller parameters.</source>
+        <translation>输入自定义的 PyInstaller 参数。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="683"/>
+        <source>Start conversion</source>
+        <translation>开始转换</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="685"/>
+        <source>Start conversion of the selected Python script to EXE.</source>
+        <translation>开始将所选 Python 脚本转换为 EXE。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="690"/>
+        <source>Cancel conversion</source>
+        <translation>取消转换</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="692"/>
+        <source>Cancel a conversion task in progress.</source>
+        <translation>取消正在进行的转换任务。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="708"/>
+        <source>Script added: {path}</source>
+        <translation>已添加脚本: {path}</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="713"/>
+        <source>Select Python file</source>
+        <translation>选择 Python 文件</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="720"/>
+        <source>Script added: {script_path}</source>
+        <translation>已添加脚本: {script_path}</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="727"/>
+        <source>Select output directory</source>
+        <translation>选择输出目录</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="733"/>
+        <source>Select Icon file</source>
+        <translation>选择图标文件</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="743"/>
+        <source>Script removed: {path}</source>
+        <translation>已移除脚本: {path}</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="906"/>
+        <source>Warning</source>
+        <translation>警告</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="749"/>
+        <source>Please select at least one Python script first.</source>
+        <translation>请先选择至少一个 Python 脚本。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="762"/>
+        <source>The file version number format is incorrect. It should be X.X.X.X (such as 1.0.0.0).</source>
+        <translation>文件版本号格式不正确，应为 X.X.X.X (如 1.0.0.0)。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="768"/>
+        <source>Start converting...</source>
+        <translation>开始转换...</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="770"/>
+        <source>Converting...</source>
+        <translation>转换中...</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="807"/>
+        <source>Cancellation of the conversion task has been requested.</source>
+        <translation>已请求取消转换任务。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="808"/>
+        <source>Cancel conversion...</source>
+        <translation>取消转换...</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="813"/>
+        <source>Conversion successful! EXE file is located at: {exe_path} (size: {exe_size} KB)</source>
+        <translation>转换成功! EXE 文件位于: {exe_path} (大小: {exe_size} KB)</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="816"/>
+        <source>Conversion successful! File: {exe_path} ({exe_size} KB)</source>
+        <translation>转换成功! 文件: {exe_path} ({exe_size} KB)</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="837"/>
+        <source>Conversion completed.</source>
+        <translation>转换完成。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="906"/>
+        <source>Log file does not exist.</source>
+        <translation>日志文件不存在。</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="924"/>
+        <source>Waiting...</source>
+        <translation>等待中...</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="929"/>
+        <source>Conversion log snippet for this task</source>
+        <translation type="obsolete">本任务的转换日志片段</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="584"/>
+        <source>Basic Settings</source>
+        <translation>基本设置</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="588"/>
+        <source>Conversion Mode:</source>
+        <translation>转换模式:</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="612"/>
+        <source>EXE Information</source>
+        <translation>EXE 信息</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="545"/>
+        <source>File</source>
+        <translation>文件</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="659"/>
+        <source>Hidden imported modules, separated by commas</source>
+        <translation>隐藏导入的模块，逗号分隔</translation>
+    </message>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="660"/>
+        <source>Enter the name of the module that needs to be added (separate multiple with commas).</source>
+        <translation>输入需要隐藏导入的模块名称（多个用逗号分隔）。</translation>
+    </message>
+</context>
+<context>
+    <name>ManualDialog</name>
+    <message>
+        <location filename="../PythonEXE_Maker.py" line="322"/>
+        <source>Manual</source>
+        <translation>使用说明</translation>
+    </message>
+</context>
+</TS>
